Fix relative path resolution (#10)
This change resolves two bugs in 20.0.0:

the nuoca script could not execute the python code from an unexpected location, leading to the following error:
```$/usr/bin/nuoca python: can't open file '/usr/pynuoca/nuoca.py': [Errno 2] No such file or directory```

the configuration script was hardcoded to a non-existent location. Fix it to be in the `$NUOCA_HOME/etc` directory. Also add the example config file to the pypi package.